various usability updates
* deadband added to prevent oscillation
* manual control added to improve flexibilty
* area filtering added for image processing
* bug fixed in velByte code where occasionally the result type would be
numpy float64

+++ src/cnc.py
@@ -158,7 +158,7 @@ class CNC:
         maxIntArg = signBit - 1
 
         # convert speed to integer representation
-        intVal = round(abs(v)*maxIntArg/self.maxSpeed)
+        intVal = int(round(abs(v)*maxIntArg/self.maxSpeed))
 
         # check that the speed is not too high
         if intVal > maxIntArg:
